+++ test/unit/utility/test_lxw_utf8_strlen.c
@@ -13,11 +13,11 @@
 // Test lxw_utf8_strlen().
 CTEST(utility, lxw_utf8_strlen) {
 
-    ASSERT_EQUAL(0,  lxw_utf8_strlen(""));
-    ASSERT_EQUAL(3,  lxw_utf8_strlen("Foo"));
-    ASSERT_EQUAL(4,  lxw_utf8_strlen("café"));
-    ASSERT_EQUAL(4,  lxw_utf8_strlen("cake"));
-    ASSERT_EQUAL(21, lxw_utf8_strlen("Это фраза на русском!"));
+    ASSERT_EQUAL(0,  (long)lxw_utf8_strlen(""));
+    ASSERT_EQUAL(3,  (long)lxw_utf8_strlen("Foo"));
+    ASSERT_EQUAL(4,  (long)lxw_utf8_strlen("café"));
+    ASSERT_EQUAL(4,  (long)lxw_utf8_strlen("cake"));
+    ASSERT_EQUAL(21, (long)lxw_utf8_strlen("Это фраза на русском!"));
 
 }
